+++ juniper/src/integrations/uuid.rs
@@ -19,8 +19,10 @@ where
         Value::scalar(self.to_string())
     }
 
-    fn from_input_value(v: &InputValue) -> Option<Uuid> {
-        v.as_string_value().and_then(|s| Uuid::parse_str(s).ok())
+    fn from_input_value(v: &InputValue) -> Result<Uuid, String> {
+        v.as_string_value()
+            .ok_or_else(|| format!("Expected `String`, found: {}", v))
+            .and_then(|s| Uuid::parse_str(s).map_err(|e| format!("Failed to parse `Uuid`: {}", e)))
     }
 
     fn from_str<'a>(value: ScalarToken<'a>) -> ParseScalarResult<'a, S> {