+++ juniper_codegen/src/graphql_object/mod.rs
@@ -486,6 +486,7 @@ impl Definition<Query> {
 
         let (impl_generics, where_clause) = self.impl_generics(false);
         let ty = &self.ty;
+        let ty_name = ty.to_token_stream().to_string();
 
         let name = &self.name;
 
@@ -493,17 +494,20 @@ impl Definition<Query> {
             .fields
             .iter()
             .filter_map(|f| f.method_resolve_field_tokens(scalar, None));
-        let async_fields_panic = {
+        let async_fields_err = {
             let names = self
                 .fields
                 .iter()
                 .filter_map(|f| f.is_async.then(|| f.name.as_str()))
                 .collect::<Vec<_>>();
             (!names.is_empty()).then(|| {
-                field::Definition::method_resolve_field_panic_async_field_tokens(&names, scalar)
+                field::Definition::method_resolve_field_err_async_field_tokens(
+                    &names, scalar, &ty_name,
+                )
             })
         };
-        let no_field_panic = field::Definition::method_resolve_field_panic_no_field_tokens(scalar);
+        let no_field_err =
+            field::Definition::method_resolve_field_err_no_field_tokens(scalar, &ty_name);
 
         quote! {
             #[allow(deprecated)]
@@ -526,8 +530,8 @@ impl Definition<Query> {
                 ) -> ::juniper::ExecutionResult<#scalar> {
                     match field {
                         #( #fields_resolvers )*
-                        #async_fields_panic
-                        _ => #no_field_panic,
+                        #async_fields_err
+                        _ => #no_field_err,
                     }
                 }
 
@@ -553,12 +557,14 @@ impl Definition<Query> {
 
         let (impl_generics, where_clause) = self.impl_generics(true);
         let ty = &self.ty;
+        let ty_name = ty.to_token_stream().to_string();
 
         let fields_resolvers = self
             .fields
             .iter()
             .map(|f| f.method_resolve_field_async_tokens(scalar, None));
-        let no_field_panic = field::Definition::method_resolve_field_panic_no_field_tokens(scalar);
+        let no_field_err =
+            field::Definition::method_resolve_field_err_no_field_tokens(scalar, &ty_name);
 
         quote! {
             #[allow(deprecated, non_snake_case)]
@@ -574,7 +580,7 @@ impl Definition<Query> {
                 ) -> ::juniper::BoxFuture<'b, ::juniper::ExecutionResult<#scalar>> {
                     match field {
                         #( #fields_resolvers )*
-                        _ => #no_field_panic,
+                        _ => Box::pin(async move { #no_field_err }),
                     }
                 }
             }