+++ juniper_codegen/src/graphql_union/mod.rs
@@ -558,11 +558,11 @@ impl Definition {
                 ) -> ::juniper::ExecutionResult<#scalar> {
                     let context = executor.context();
                     #( #variant_resolvers )*
-                    panic!(
+                    return Err(::juniper::FieldError::from(format!(
                         "Concrete type `{}` is not handled by instance \
                          resolvers on GraphQL union `{}`",
                         type_name, #name,
-                    );
+                    )));
                 }
             }
         }
@@ -600,11 +600,11 @@ impl Definition {
                 ) -> ::juniper::BoxFuture<'b, ::juniper::ExecutionResult<#scalar>> {
                     let context = executor.context();
                     #( #variant_async_resolvers )*
-                    panic!(
+                    return ::juniper::macros::helper::err_fut(format!(
                         "Concrete type `{}` is not handled by instance \
                          resolvers on GraphQL union `{}`",
                         type_name, #name,
-                    );
+                    ));
                 }
             }
         }
@@ -670,11 +670,14 @@ impl VariantDefinition {
     #[must_use]
     fn method_resolve_into_type_tokens(&self, scalar: &scalar::Type) -> TokenStream {
         let ty = &self.ty;
+        let ty_name = ty.to_token_stream().to_string();
         let expr = &self.resolver_code;
         let resolving_code = gen::sync_resolving_code();
 
         quote! {
-            if type_name == <#ty as ::juniper::GraphQLType<#scalar>>::name(info).unwrap() {
+            if type_name == <#ty as ::juniper::GraphQLType<#scalar>>::name(info)
+                .ok_or_else(|| ::juniper::macros::helper::err_unnamed_type(#ty_name))?
+            {
                 let res = { #expr };
                 return #resolving_code;
             }
@@ -690,13 +693,19 @@ impl VariantDefinition {
     #[must_use]
     fn method_resolve_into_type_async_tokens(&self, scalar: &scalar::Type) -> TokenStream {
         let ty = &self.ty;
+        let ty_name = ty.to_token_stream().to_string();
         let expr = &self.resolver_code;
         let resolving_code = gen::async_resolving_code(None);
 
         quote! {
-            if type_name == <#ty as ::juniper::GraphQLType<#scalar>>::name(info).unwrap() {
-                let fut = ::juniper::futures::future::ready({ #expr });
-                return #resolving_code;
+            match <#ty as ::juniper::GraphQLType<#scalar>>::name(info) {
+                Some(name) => {
+                    if type_name == name {
+                        let fut = ::juniper::futures::future::ready({ #expr });
+                        return #resolving_code;
+                    }
+                }
+                None => return ::juniper::macros::helper::err_unnamed_type_fut(#ty_name),
             }
         }
     }